requirements.txt
aerich==0.2.2
aiomysql==0.0.20
aiosqlite==0.15.0
anyio==1.4.0
async-generator==1.10
asyncclick==7.0.9
asyncpg==0.21.0
bcrypt==3.2.0
cffi==1.14.2
click==7.1.2
cryptography==3.1
ecdsa==0.14.1
fastapi==0.61.1
h11==0.9.0
httptools==0.1.1
idna==2.10
iso8601==0.1.12
passlib==1.7.2
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
PyMySQL==0.9.2
PyPika==0.39.1
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.6
six==1.15.0
sniffio==1.1.0
starlette==0.13.6
tortoise-orm==0.16.14
typing-extensions==3.7.4.3
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
